depart into two file
