+++ src/operator/tensor/index_add_backward.cc
@@ -27,60 +27,6 @@
 namespace mxnet {
 namespace op {
 
-template<typename DType, typename VType, int NDim>
-struct IndexAddForwardCPUKernel {
-  MSHADOW_XINLINE static void Map(size_t i, DType* out,
-                                  const VType* val,
-                                  const mshadow::Shape<NDim> a_tail_shape,
-                                  const mshadow::Shape<NDim> a_pre_stride,
-                                  const mshadow::Shape<NDim> val_stride,
-                                  const mshadow::Shape<NDim> val_shape,
-                                  const size_t a_tail_size, const int ind_num,
-                                  const int ind_ndim, const int* ind_vec) {
-    size_t id = 0;
-    for (int dim = 0; dim < ind_ndim; ++dim) {
-      id += a_pre_stride[dim] * ind_vec[dim * ind_num + i];
-    }
-    id *= a_tail_size;
-    #pragma omp parallel for
-    for (int _i = 0; _i < a_tail_size; ++_i) {
-      mshadow::Shape<NDim> a_tail_id = mxnet_op::unravel(_i, a_tail_shape);
-      mshadow::Shape<NDim> val_id;
-      for (int _j = 0; _j < NDim; ++_j) {
-        val_id[_j] = (val_shape[_j] == 1) ? 0 : a_tail_id[_j];
-      }
-      val_id[ind_ndim - 1] = (val_shape[ind_ndim - 1] == 1) ? 0 : i;
-      size_t val_dest = mxnet_op::dot(val_id, val_stride);
-      #pragma omp critical
-      {
-        out[id + _i] += static_cast<DType>(val[val_dest]);
-      }
-      // #pragma omp reduction(+:out[id + _i])
-      // KERNEL_ASSIGN(out[id + _i], req, out[id + _i] + static_cast<DType>(val[val_dest]));
-    }
-  }
-};
-
-template<typename xpu, typename DType, typename VType, int NDim>
-void IndexAddForwardCalc(mshadow::Stream<xpu> *s,
-                         const int ind_num, DType* out,
-                        const VType* val,
-                        const mshadow::Shape<NDim>& a_tail_shape,
-                        const mshadow::Shape<NDim>& a_pre_stride,
-                        const mshadow::Shape<NDim>& val_stride,
-                        const mshadow::Shape<NDim>& val_shape,
-                        const size_t a_tail_size,
-                        const int ind_ndim, const int* ind_vec) {
-  using namespace mxnet_op;
-  using namespace mshadow;
-  Kernel<IndexAddForwardCPUKernel<DType, VType, NDim>, xpu>::Launch(
-                                             s, ind_num, out, val,
-                                             a_tail_shape, a_pre_stride,
-                                             val_stride, val_shape,
-                                             a_tail_size, ind_num,
-                                             ind_ndim, ind_vec);
-}
-
 template<typename xpu, typename DType, typename OType, int NDim>
 void IndexAddOpBackwardACalc(mshadow::Stream<xpu> *s,
                              DType* grad_a, const OType* ograd,
@@ -141,26 +87,6 @@ void IndexAddOpBackwardValCalc(mshadow::Stream<xpu> *s,
     s, ind_num, grad_val, ograd, ograd_tail_shape, ograd_pre_stride,
     val_stride, val_shape, tail_size, ind_num, ind_ndim, ind_vec);
 }
-
-DMLC_REGISTER_PARAMETER(IndexModifyParam);
-
-NNVM_REGISTER_OP(_npx_index_add)
-.describe(R"code(This operators implements the "+=" mimic function.
-)code" ADD_FILELINE)
-.set_attr_parser(ParamParser<IndexModifyParam>)
-.set_num_inputs(2)
-.set_num_outputs(1)
-.set_attr<nnvm::FListInputNames>("FListInputNames",
-  [](const NodeAttrs& attrs) {
-    return std::vector<std::string>{"a", "val"};
-  })
-.set_attr<mxnet::FInferShape>("FInferShape", IndexModifyOpShape)
-.set_attr<nnvm::FInferType>("FInferType", IndexModifyOpType)
-.set_attr<FCompute>("FCompute<cpu>", IndexAddOpForward<cpu>)
-.set_attr<nnvm::FGradient>("FGradient", ElemwiseGradUseIn{"_backward_index_add"})
-.add_argument("a", "NDArray-or-Symbol", "Input ndarray")
-.add_argument("val", "NDArray-or-Symbol", "Input ndarray")
-.add_arguments(IndexModifyParam::__FIELDS__());
 
 NNVM_REGISTER_OP(_backward_index_add)
 .set_num_inputs(3)
